variational-solver: move nonlinear block to fenics-adjoint
The the firedrake implementation will be in firedrake.adjoint.
