+++ dolfin_adjoint_common/blocks/__init__.py
@@ -1,7 +1,6 @@
 from .assembly import AssembleBlock
 from .solving import GenericSolveBlock, SolveVarFormBlock, SolveLinearSystemBlock
 from .projection import ProjectBlock
-from .variational_solver import NonlinearVariationalSolveBlock
 from .function import FunctionAssignBlock
 from .function_assigner import FunctionAssignerBlock
 from .mesh import ALEMoveBlock, BoundaryMeshBlock
