+++ fenics_adjoint/blocks/__init__.py
@@ -1,5 +1,5 @@
 from .common import *
-from .variational_solver import LinearVariationalSolveBlock
+from .variational_solver import LinearVariationalSolveBlock, NonlinearVariationalSolveBlock
 from .krylov_solver import KrylovSolveBlock, KrylovSolveBlockHelper
 from .lu_solver import LUSolveBlock, LUSolveBlockHelper
 from .petsc_krylov_solver import PETScKrylovSolveBlock, PETScKrylovSolveBlockHelper
